Collapse RelationshipType into Relationship
